+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to the **Socarel** project will be documented in this file.
 - Iterate subtrees from any initial node using `Tree::iterators_at()`.
 - Children iterator.
 - New function to find paths, `Tree::find_path`.
+- Create a trait for Tree IDs, `TreeIdentifier`, similar to NodeContent.
 
 ### Changes